initial gpu impl

+++ include/connections/linear_algebra/gpu/cuda/activations.cuh
@@ -1,12 +1,12 @@
 #pragma once
 
 namespace cntns {
-__global__ void logsig_kernel(double const* input, double* output, int size);
-__global__ void logsig_derivative_kernel(double const* input, double* output, int size);
+__global__ void logsig_kernel(float const* input, float* output, int size);
+__global__ void logsig_derivative_kernel(float const* input, float* output, int size);
 
-__global__ void relu_kernel(double const* input, double* output, int size);
-__global__ void relu_derivative_kernel(double const* input, double* output, int size);
+__global__ void relu_kernel(float const* input, float* output, int size);
+__global__ void relu_derivative_kernel(float const* input, float* output, int size);
 
-__global__ void softmax_kernel(double const* input, double* output, int size);
+__global__ void softmax_kernel(float const* input, float* output, int size);
 }
