+++ include/connections/linear_algebra/gpu/cuda/loss.cuh
@@ -1,7 +1,7 @@
 #pragma once
 
 namespace cntns {
-__global__ void quadratic_cost_kernel(double const* correct, double const* result, double* loss, int size);
+__global__ void quadratic_cost_kernel(float const* correct, float const* result, float* loss, int size);
 
-__global__ void cross_entropy(double const* correct, double const* result, double* loss, int size);
+__global__ void cross_entropy(float const* correct, float const* result, float* loss, int size);
 }
