+++ include/connections/linear_algebra/gpu/cuda/vector.cuh
@@ -2,15 +2,15 @@
 #pragma once
 
 namespace cntns {
-__global__ void vector_randomize_kernel(double* data, double min, double max, int size);
-__global__ void vector_fill_kernel(double* data, double value, int size);
-__global__ void vector_reset_kernel(double* data, int size);
+__global__ void vector_randomize_kernel(float* data, float min, float max, int size);
+__global__ void vector_fill_kernel(float* data, float value, int size);
+__global__ void vector_reset_kernel(float* data, int size);
 
-__global__ void max_element_kernel(double* input, int* maxIndex, int size);
+__global__ void max_element_kernel(float* input, int* maxIndex, int size);
 
-__global__ void vector_add_kernel(double const* vecA, double const* vecB, double* out, int size);
-__global__ void vector_sub_kernel(double const* vecA, double const* vecB, double* out, int size);
-__global__ void vector_mul_kernel(double const* vecA, double const* vecB, double* out, int size);
-__global__ void vector_scalar_kernel(double const* vecA, double vecB, double* out, int size);
-__global__ void outer_product_kernel(double const* lhs, double const* rhs, double* result, int lhs_size, int rhs_size);
+__global__ void vector_add_kernel(float const* vecA, float const* vecB, float* out, int size);
+__global__ void vector_sub_kernel(float const* vecA, float const* vecB, float* out, int size);
+__global__ void vector_mul_kernel(float const* vecA, float const* vecB, float* out, int size);
+__global__ void vector_scalar_kernel(float const* vecA, float vecB, float* out, int size);
+__global__ void outer_product_kernel(float const* lhs, float const* rhs, float* result, int lhs_size, int rhs_size);
 }
