+++ src/matrix.cu
@@ -35,7 +35,7 @@ __device__ auto index(unsigned int row, unsigned int col, int cols) -> unsigned 
  * @param cols 
  * @return __global__ 
  */
-__global__ void mat_randomize_kernel(double* data, double min, double max, int rows, int cols) {
+__global__ void mat_randomize_kernel(float* data, float min, float max, int rows, int cols) {
   unsigned int row = blockIdx.y * blockDim.y + threadIdx.y;
   unsigned int col = blockIdx.x * blockDim.x + threadIdx.x;
 
@@ -57,7 +57,7 @@ __global__ void mat_randomize_kernel(double* data, double min, double max, int r
  * @param cols 
  * @return __global__ 
  */
-__global__ void mat_fill_kernel(double* data, double value, int rows, int cols) {
+__global__ void mat_fill_kernel(float* data, float value, int rows, int cols) {
   unsigned int row = blockIdx.y * blockDim.y + threadIdx.y;
   unsigned int col = blockIdx.x * blockDim.x + threadIdx.x;
 
@@ -76,7 +76,7 @@ __global__ void mat_fill_kernel(double* data, double value, int rows, int cols) 
  * @param cols 
  * @return __global__ 
  */
-__global__ void mat_add_kernel(double const* A, double const* B, double* out, int rows, int cols) {
+__global__ void mat_add_kernel(float const* A, float const* B, float* out, int rows, int cols) {
   unsigned int row = blockIdx.y * blockDim.y + threadIdx.y;
   unsigned int col = blockIdx.x * blockDim.x + threadIdx.x;
 
@@ -94,7 +94,7 @@ __global__ void mat_add_kernel(double const* A, double const* B, double* out, in
  * @param inCols 
  * @return __global__ 
  */
-__global__ void transpose_kernel(double const* input, double* output, int inRows, int inCols) {
+__global__ void transpose_kernel(float const* input, float* output, int inRows, int inCols) {
   unsigned int row = blockIdx.y * blockDim.y + threadIdx.y;
   unsigned int col = blockIdx.x * blockDim.x + threadIdx.x;
 
@@ -114,7 +114,7 @@ __global__ void transpose_kernel(double const* input, double* output, int inRows
  * @param cols 
  * @return __global__ 
  */
-__global__ void mat_outer_product_add_kernel(double * A, double const* lhs, double const* rhs, int rows, int cols) {
+__global__ void mat_outer_product_add_kernel(float * A, float const* lhs, float const* rhs, int rows, int cols) {
   unsigned int row = blockIdx.y * blockDim.y + threadIdx.y;
   unsigned int col = blockIdx.x * blockDim.x + threadIdx.x;
 
@@ -133,7 +133,7 @@ __global__ void mat_outer_product_add_kernel(double * A, double const* lhs, doub
  * @param cols 
  * @return __global__ 
  */
-__global__ void mat_sub_kernel(double const* A, double const* B, double* out, int rows, int cols) {
+__global__ void mat_sub_kernel(float const* A, float const* B, float* out, int rows, int cols) {
   unsigned int row = blockIdx.y * blockDim.y + threadIdx.y;
   unsigned int col = blockIdx.x * blockDim.x + threadIdx.x;
 
@@ -152,7 +152,7 @@ __global__ void mat_sub_kernel(double const* A, double const* B, double* out, in
  * @param cols 
  * @return __global__ 
  */
-__global__ void mat_mul_kernel(double const* A, double const* B, double* out, int rows, int cols) {
+__global__ void mat_mul_kernel(float const* A, float const* B, float* out, int rows, int cols) {
   unsigned int row = blockIdx.y * blockDim.y + threadIdx.y;
   unsigned int col = blockIdx.x * blockDim.x + threadIdx.x;
 
@@ -172,7 +172,7 @@ __global__ void mat_mul_kernel(double const* A, double const* B, double* out, in
  * @param colsB 
  * @return __global__ 
  */
-__global__ void mat_scalar_kernel(double const* A, double B, double* out, int rows, int cols) {
+__global__ void mat_scalar_kernel(float const* A, float B, float* out, int rows, int cols) {
   unsigned int row = blockIdx.y * blockDim.y + threadIdx.y;
   unsigned int col = blockIdx.x * blockDim.x + threadIdx.x;
 
@@ -190,7 +190,7 @@ __global__ void mat_scalar_kernel(double const* A, double B, double* out, int ro
  * @param cols 
  * @return __global__ 
  */
-__global__ void mat_scalar_eq_kernel(double* A, double B, int rows, int cols) {
+__global__ void mat_scalar_eq_kernel(float* A, float B, int rows, int cols) {
   unsigned int row = blockIdx.y * blockDim.y + threadIdx.y;
   unsigned int col = blockIdx.x * blockDim.x + threadIdx.x;
 
@@ -209,11 +209,11 @@ __global__ void mat_scalar_eq_kernel(double* A, double B, int rows, int cols) {
  * @param cols 
  * @return __global__ 
  */
-__global__ void mat_vec_mul_kernel(double const* A, double const* B, double* out, int rows, int cols) {
+__global__ void mat_vec_mul_kernel(float const* A, float const* B, float* out, int rows, int cols) {
   unsigned int row = blockIdx.x * blockDim.x + threadIdx.x;
   
   if (row < rows) {
-    double sum = 0.0F;
+    float sum = 0.0F;
     for (int i = 0; i < cols; ++i) {
       sum +=  A[index(row, i, cols)] * B[i];
     }
@@ -231,11 +231,11 @@ __global__ void mat_vec_mul_kernel(double const* A, double const* B, double* out
  * @param cols 
  * @return __global__ 
  */
-__global__ void mat_transpose_vec_mul_kernel(double const* A, double const* B, double* out, int rows, int cols) {
+__global__ void mat_transpose_vec_mul_kernel(float const* A, float const* B, float* out, int rows, int cols) {
   unsigned int row = blockIdx.x * blockDim.x + threadIdx.x;
   
   if (row < cols) {
-    double sum = 0.0F;
+    float sum = 0.0F;
     for (int i = 0; i < rows; ++i) {
       sum +=  A[index(i, row, cols)] * B[i];
     }
